requirements.txt
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
matplotlib==3.8.2
seaborn==0.13.2
pyyaml==6.0.1
pytest==8.0.0
requests==2.31.0
pandera==0.18.0
python-multipart==0.0.9
